Correct help text and default for server URL arg
Not sure why I thought the InfluxDB 2.x client library accepted a bare host/port for the server, in addition to a full URL. It does not, as reported in issue #49.  The 1.x client did, but it didn't take a URL for server, anyway.

This corrects the help text to indicate the URL argument specifies a... URL, and corrects the default to be something that could actually work.