Type hints for status data return values
Implement type hints for the data returned from status_data, while using the same base data for the implementations of status_field_names and status_field_types.

This is complicated a little by the fact that the field names returned for sequences in the returned data are slightly different from those in status_field_names, for reasons that I'm sure made perfect sense to me at the time.

This is for issue #57