Make Docker image work for 32-bit ARMv7
This should allow for a working image if built manually, but does not add it to the multi-arch package built for this project's package repository.

As discussed in issue #72, needed to work around some shared library issue between the binary wheel for one of the prerequisites published to PyPI and the base Docker image.