Add a function to get current sleep mode config
Bring the core module functionality back into parity with dish_control,py, to which I had added the ability to both get and set the sleep mode config, not just set it.

Also, some mostly relevant README updates that I meant to check in a long time ago.

This is for issue #86.