Document details of the obstruction map image
Attempt to provide a technically accurate description of how observation direction maps to pixel coordinates, and what exactly the color represents. This is based on empirical observation of the data from my dish, but hopefully will apply even for dishes located in vastly different parts of the globe.

Related to the question asked in issue #85.