Add ability to reset the obstruction map data
Dish firmware support for this was added a week or 2 ago.

I don't expect this to be particularly useful unless you want to see the data fill  in over time without having to reboot your dish, and it will probably result in increased dropped packets due to obstructions until the data has filled in again.