Add a script to check and trigger software update
This script polls the dish status for whether or not a software update is pending and can optionally reboot the dish in order to initiate the install if there is one pending.

It's possible there is more to initiating software updates for some case than what this script does, but this appears to work for triggering the install without waiting for the (typically overnight?) automatic install.

It currently has the same basic periodic loop control as most of the other scripts, which is probably not that useful for this type of script. This may change in the near future.

This is for issue #89