Implement cron-like scheduling for update checker
Implement a loop option to run at specific wall clock times, based on the same schedule string format that cron uses for crontab entries. Default behavior is to use the system local time zone, but this can be overridden via command line arg.

For now, this is just for the update check script, but it may be applied to the other scripts that have periodic loop options.

This adds some new Python package dependencies, but they are imported in a way that they won't break unless the new functionality that requires them is used.

Also, change the short option for the --target option to -g, which is what most of the other scripts use, instead of -e, which is only used by the obstruction map script.

This is related to issue #89