Add warning if updates disabled
Check if the software update state is "disabled" and report that if it is. Not sure if that's even something that is possible to happen, or if it's relevant to the other ways this script can detect software update pending, but it seems worth noting.

This is probably the last functional change for issue #89 before I consider that done, pending further testing.