Update the versions of all actions used
Hopefully, this will address the warnings that show up every time this workflow is run, given that the warnings look like things not done directly here.

Also, enable the 32-bit ARM image build and publish, because that has been sitting in my tree since forever. If this causes a problem, I'll back this part out.