Add support for power consumption data
Add power consumption data into both bulk history and history stats reporting. Note that this functionality is not supported on all user terminal (dish) hardware versions.

In the core module, for bulk history, this is added as a new field in the bulk history return dict. For history stats, this is a added as a newly added dict in the tuple returned from the history_stats.

In the command line tools, the existing bulk_history data mode will now return the power history as a new field, and the newly added power data mode will return a few statistics computed over the sample period.

This is for issue #94.