Work around recent firmware behavior change
In recent dish firmware versions, the ping latency always reports as 0 in the status grpc response. It's correct in the history response, so switch to using that for the latency field of the status data group, at least for now. I'm not sure whether or not this affects the ping drop number, too, since that is normally 0, so to be safe, that will be taken from history, too.

This should address issue #100.

This may just be a bug in the firmware that may get fixed in the future. If so, I will probably back this change out at that time, as it results in an extra grpc request being issued.