Add a note about venv to install instructions
As with any Python project that requires modules to be installed via pip, use of a virtual environment is a good idea for this project. I hadn't mentioned that in an effort to keep things simple and because I figured people who cared about such things already knew about venv and how to use it.

As mentioned in issue #99, though, some OS distributions have started marking the system Python packages as owned by the system and will throw an error if you try to install additional modules there using pip. venv instructions vary a little by system, though, so rather than dictate that people have to use it in a certain way and complicate the install instructions to boot, there is now a note about it being recommended and sometimes required along with a link to usage instructions.