Add ability to run periodically to dish_control
Add the same loop control options as dish_check_update.py has. This is to facilitate automation, where a command needs to be run repeatedly and/or at a specific time.

This does further complicate a script I was hoping to keep relatively simple, but I'm pretty sure nobody cares about that other than me.

Also, update the README to reflect GPS control ability (such that it is).

Vaguely related to issue #101.