Back out workaround for latency firmware bug
Go back to pulling ping latency and drop rate from status instead of history when pulling status info, removing the need for an extra query of history data.

This removes the workaround for a firmware bug that was breaking that data in the status info (see issue #100). That bug appears to have been fixed in the firmware well over a month ago, so all deployed dishes should have gotten the fix by now.