Add compatibility fallback for package change
If resolving grpc protocol modules fails against the newly changed package name, try the old one.

This is a little messy, because yagrc's lazy import mechanism isn't meant for this level of flexibility. As such, if the lazy import mechanism fails, the fallback uses the non-lazy pattern. This can be removed once the firmware with the new package name has fully rolled out; otherwise, there wouldn't be much point in using the lazy import mode in yagrc.